feat: add tests

+++ later42/views/api_token.py
@@ -1,13 +1,13 @@
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect
 from rest_framework.authtoken.models import Token
 
 
+@login_required
 def create(request):
     token = Token.objects.filter(user=request.user)
-    try:
+    if len(token) > 0:
         token.delete()
-    except:
-        pass
     token = Token.objects.create(user=request.user)
     token.save()
     return redirect('profile')
